Fetch schemas and store objectsArray in Redux state
