+++ src/components/NewDraftWizard/WizardComponents/WizardObjectIndex.js
@@ -1,5 +1,5 @@
 //@flow
-import React, { useState } from "react"
+import React, { useState, useEffect } from "react"
 
 import MuiAccordion from "@material-ui/core/Accordion"
 import MuiAccordionDetails from "@material-ui/core/AccordionDetails"
@@ -15,12 +15,14 @@ import { useDispatch, useSelector } from "react-redux"
 
 import WizardAlert from "./WizardAlert"
 
-import { ObjectSubmissionTypes, ObjectSubmissionsArray, getObjectsArray } from "constants/wizardObject"
+import { ObjectSubmissionTypes, ObjectSubmissionsArray, ObjectTypes } from "constants/wizardObject"
 import { resetDraftStatus } from "features/draftStatusSlice"
 import { setFocus } from "features/focusSlice"
+import { setObjectsArray } from "features/objectsArraySlice"
 import { resetCurrentObject } from "features/wizardCurrentObjectSlice"
 import { setObjectType } from "features/wizardObjectTypeSlice"
 import { setSubmissionType } from "features/wizardSubmissionTypeSlice"
+import schemaAPIService from "services/schemaAPI"
 
 const useStyles = makeStyles(theme => ({
   index: {
@@ -216,11 +218,11 @@ const WizardObjectIndex = (): React$Element<any> => {
   const classes = useStyles()
   const dispatch = useDispatch()
 
-  const objectsArray = getObjectsArray()
-
   const [expandedObjectType, setExpandedObjectType] = useState("")
   const [clickedSubmissionType, setClickedSubmissionType] = useState("")
   const [cancelFormOpen, setCancelFormOpen] = useState(false)
+
+  const objectsArray = useSelector(state => state.objectsArray)
   const currentObjectType = useSelector(state => state.objectType)
   const currentSubmissionType = useSelector(state => state.submissionType)
   const draftStatus = useSelector(state => state.draftStatus)
@@ -231,6 +233,44 @@ const WizardObjectIndex = (): React$Element<any> => {
   const draftObjects = folder.drafts
     ?.map(draft => draft.schema)
     .reduce((acc, val) => ((acc[val] = (acc[val] || 0) + 1), acc), {})
+
+  // Fetch array of schemas from backend and store it in frontend
+  // Fetch only if the initial array is empty
+  // if there is any errors while fetching, it will return a manually created ObjectsArray instead
+  useEffect(() => {
+    if (objectsArray.length === 0) {
+      let isMounted = true
+      const getSchemas = async () => {
+        const response = await schemaAPIService.getAllSchemas()
+
+        if (isMounted) {
+          if (response.ok) {
+            const schemas = response.data
+              .filter(schema => schema.title !== "Project" && schema.title !== "Submission")
+              .map(schema => schema.title.toLowerCase())
+            dispatch(setObjectsArray(schemas))
+          } else {
+            dispatch(
+              setObjectsArray([
+                ObjectTypes.study,
+                ObjectTypes.sample,
+                ObjectTypes.experiment,
+                ObjectTypes.run,
+                ObjectTypes.analysis,
+                ObjectTypes.dac,
+                ObjectTypes.policy,
+                ObjectTypes.dataset,
+              ])
+            )
+          }
+        }
+      }
+      getSchemas()
+      return () => {
+        isMounted = false
+      }
+    }
+  }, [])
 
   const handlePanelChange = panel => (event, newExpanded) => {
     setExpandedObjectType(newExpanded ? panel : false)
