+++ src/constants/wizardObject.js
@@ -1,7 +1,3 @@
-import { useEffect, useState } from "react"
-
-import schemaAPIService from "services/schemaAPI"
-
 export const ObjectTypes = {
   study: "study",
   sample: "sample",
@@ -11,41 +7,6 @@ export const ObjectTypes = {
   dac: "dac",
   policy: "policy",
   dataset: "dataset",
-}
-
-export const fetchSchemas = async () => {
-  const response = await schemaAPIService.getAllSchemas()
-  if (response.ok) {
-    const schemas = response.data
-      .filter(schema => schema.title !== "Project" && schema.title !== "Submission")
-      .map(schema => schema.title.toLowerCase())
-    return schemas
-  } else {
-    return [
-      ObjectTypes.study,
-      ObjectTypes.sample,
-      ObjectTypes.experiment,
-      ObjectTypes.run,
-      ObjectTypes.analysis,
-      ObjectTypes.dac,
-      ObjectTypes.policy,
-      ObjectTypes.dataset,
-    ]
-  }
-}
-
-export const getObjectsArray = () => {
-  const [objectsArray, setObjectsArray] = useState([])
-
-  useEffect(() => {
-    const getSchemas = async () => {
-      const schemas = await fetchSchemas()
-      setObjectsArray(schemas)
-    }
-    getSchemas()
-  }, [objectsArray.length])
-
-  return objectsArray
 }
 
 export const ObjectStatus = {
